requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
markdown==3.5.1
bleach==6.1.0
python-dateutil==2.8.2
gunicorn==21.2.0
PyMySQL==1.1.0
cryptography==41.0.7
SQLAlchemy==2.0.23
